+++ Components/Core/Conditions/Categories/Talents.lua
@@ -31,20 +31,21 @@ ConditionCategory:RegisterCondition(0.2,  "CLASS2", {
 	bitFlagTitle = L["CONDITIONPANEL_BITFLAGS_CHOOSECLASS"],
 	bitFlags = (function()
 		local t = {}
-		for i = 1, MAX_CLASSES do
-			local token = CLASS_SORT_ORDER[i]
-			local name = LOCALIZED_CLASS_NAMES_MALE[token]
-			t[i] = {
-				order = i,
-				text = PLAYER_CLASS_NO_SPEC:format(RAID_CLASS_COLORS[token].colorStr, name),
-				icon = "Interface\\GLUES\\CHARACTERCREATE\\UI-CHARACTERCREATE-CLASSES",
-				tcoords = {
-					(CLASS_ICON_TCOORDS[token][1]+.02),
-					(CLASS_ICON_TCOORDS[token][2]-.02),
-					(CLASS_ICON_TCOORDS[token][3]+.02),
-					(CLASS_ICON_TCOORDS[token][4]-.02),
+		for classID = 1, TMW.GetMaxClassID() do
+			local name, token = TMW.GetClassInfo(classID)
+			if name then
+				t[classID] = {
+					order = classID,
+					text = PLAYER_CLASS_NO_SPEC:format(RAID_CLASS_COLORS[token].colorStr, name),
+					icon = "Interface\\GLUES\\CHARACTERCREATE\\UI-CHARACTERCREATE-CLASSES",
+					tcoords = {
+						(CLASS_ICON_TCOORDS[token][1]+.02),
+						(CLASS_ICON_TCOORDS[token][2]-.02),
+						(CLASS_ICON_TCOORDS[token][3]+.02),
+						(CLASS_ICON_TCOORDS[token][4]-.02),
+					}
 				}
-			}
+			end
 		end
 		return t
 	end)(),