+++ TellMeWhen.lua
@@ -33,7 +33,7 @@ if not wow_classic then
 	return
 end
 
-TELLMEWHEN_VERSION = "8.6.9"
+TELLMEWHEN_VERSION = "8.7.0"
 
 TELLMEWHEN_VERSION_MINOR = ""
 local projectVersion = "@project-version@" -- comes out like "6.2.2-21-g4e91cee"
@@ -44,11 +44,11 @@ elseif strmatch(projectVersion, "%-%d+%-") then
 end
 
 TELLMEWHEN_VERSION_FULL = TELLMEWHEN_VERSION .. " Classic " .. TELLMEWHEN_VERSION_MINOR
-TELLMEWHEN_VERSIONNUMBER = 86907 -- NEVER DECREASE THIS NUMBER (duh?).  IT IS ALSO ONLY INTERNAL (for versioning of)
+TELLMEWHEN_VERSIONNUMBER = 87001 -- NEVER DECREASE THIS NUMBER (duh?).  IT IS ALSO ONLY INTERNAL (for versioning of)
 
 TELLMEWHEN_FORCECHANGELOG = 86005 -- if the user hasn't seen the changelog until at least this version, show it to them.
 
-if TELLMEWHEN_VERSIONNUMBER > 87000 or TELLMEWHEN_VERSIONNUMBER < 86000 then
+if TELLMEWHEN_VERSIONNUMBER > 88000 or TELLMEWHEN_VERSIONNUMBER < 87000 then
 	-- safety check because i accidentally made the version number 414069 once
 	return error("TELLMEWHEN: THE VERSION NUMBER IS SCREWED UP OR MAYBE THE SAFETY LIMITS ARE WRONG")
 end
@@ -1081,37 +1081,8 @@ function TMW:PLAYER_LOGIN()
 		return
 
 	-- if the file IS required for gross functionality
-	elseif not TMW.BE then
-		local fileName = "TellMeWhen/Components/Core/Spells/Equivalencies.lua"
-
-
-		-- Ok, so this check clearly has some problems. Maybe? For years now,
-		-- i've been getting occasional reports that this isn't detecting things properly,
-		-- and that it just continually pops up no matter what people do.
-		-- So, instead of forcing a restart on people, i'm going to take out the early return and instead,
-		-- output a ton of debug information.
-		-- local classCount = 0
-		-- for k, v in pairs(TMW.C) do classCount = classCount + 1 end
-
-		-- TMW:Print("There was an issue during TMW's Initialization. A required file, " .. fileName .. " didn't seem to load." )
-		-- TMW:Print("If you haven't restarted WoW since last updating it, please do so now." )
-		-- TMW:Print("If you have restarted and this error keeps happening, please report the following information to the addon page at Curse.com (a screenshot of this would probably be easiest):" )
-		-- TMW:Print(
-		-- 	"v", TELLMEWHEN_VERSIONNUMBER, 
-		-- 	"TMW.C count", classCount,
-		-- 	"TMW.BE", TMW.BE,
-		-- 	"TMW.CNDT", TMW.CNDT, 
-		-- 	"toc v",  GetAddOnMetadata("TellMeWhen", "Version"),
-		-- 	"xcpv",  GetAddOnMetadata("TellMeWhen", "X-Curse-Packaged-Version"),
-		-- 	"dbvar", TellMeWhenDB,
-		-- 	"dbver", TellMeWhenDB and TellMeWhenDB.Version,
-		-- 	"mac?", IsMacClient(),
-		-- 	"wowb", select(2, GetBuildInfo()),
-		-- 	"L", TMW.L,
-		-- 	"ldb", LibStub("LibDataBroker-1.1") and LibStub("LibDataBroker-1.1"):GetDataObjectByName("TellMeWhen") or "noldb",
-		-- 	"types", TMW.approachTable and #(TMW.approachTable(TMW, "C", "IconType", "instances") or {}) or "noapproach"
-		-- )
-
+	elseif not LibStub("LibClassicCasterino", true) then
+		local fileName = "TellMeWhen/Lib/LibClassicCasterino/LibClassicCasterino.lua"
 
 		-- this also includes upgrading from older than 3.0 (pre-Ace3 DB settings)
 		-- GLOBALS: StaticPopupDialogs, StaticPopup_Show, EXIT_GAME, CANCEL, ForceQuit