+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v8.7.0
+* Updated the Cast condition and icon type to use LibClassicCasterino for approximations of others' spell casts.
+
+### Bug Fixes
+* Classic: Fixed the Unit Class condition's options.
+
 ## v8.6.9
 * Classic: Aura durations now account for learned talents.
 * Classic: Swing Timer now accounts for next-swing abilities.